// service/types.gen.ts
export type DistrictResponseSchema = {
    id: string;
    name: string;
};

export type CityResponseSchema = {
    id: string;
    name: string;
};

export type CbdResponseSchema = {
    id: string;
    name: string;
    addr: string | null;
};